examples/steering_app_server/requirements.txt
asn1crypto==1.0.1
bcrypt==3.1.7
cffi==1.12.3
cryptography==2.7
grpcio==1.22.0
grpcio-tools==1.22.0
lxml==4.4.1
ncclient==0.6.6
paramiko==2.6.0
protobuf==3.9.1
pycparser==2.19
PyNaCl==1.3.0
scapy==2.4.3
six==1.12.0
